Feat - change local function of get all repositories
